Add bip-0039 words
Added BIP39 word list to be able to generate mnemonic phrases as a method of encoding a radom or deterministic wallet seed in a way that's easier to for users to interpret and record and remember than binary. It is used for the mnemonic phrase in cryptographic recovery keys and in passphrases. This means there're 4,194,304 unique variables and a user could potentially have a passphrase that is 4194304 words long (This is 2048*2048). Using this format instead of binary significantly reduces the riskof errors during backup and/or recovery; a balance between security and usability.